requirements.txt
altair==5.5.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
easyocr==1.7.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.55.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.15.2
networkx==3.4.2
ninja==1.11.1.2
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.0.2
Pillow==9.5.0
protobuf==5.29.1
psutil==6.1.0
pyarrow==18.1.0
pyclipper==1.3.0.post6
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.24.0
scipy==1.10.1
setuptools==75.6.0
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tifffile==2024.9.20
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.0.114
urllib3==2.2.3
watchdog==6.0.0
streamlit-keycloak==1.1.1
kafka-python==2.0.2
